Renames the root reducer file
